# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API do W-AZAP",
    "version": "1.6.4",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor da API"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave de API gerada no painel (Webhooks e API)"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "Cookie de login do painel (somente navegador)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error occurred"
          },
          "error": {
            "type": "string",
            "example": "Detailed error info"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autenticado: chave de API ausente ou inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Proibido: acesso negado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Forbidden - Cannot access this session",
              "error": "Forbidden - Cannot access this session"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "paths": {
    "/media/{filename}": {
      "get": {
        "tags": [
          "Mídia"
        ],
        "summary": "Baixar arquivo de mídia",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conteúdo do arquivo"
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "tags": [
          "Autenticação web"
        ],
        "summary": "Cadastrar novo usuário",
        "description": "Cadastra um usuário pelo navegador. Fica desativado por padrão: só a primeira conta de uma instalação nova (que vira SUPERADMIN) pode se cadastrar sem que o administrador ative o cadastro em Configurações. Senha com no mínimo 8 caracteres; limite de 10 cadastros por IP por hora.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário cadastrado"
          }
        }
      }
    },
...
    "/autoreplies/{sessionId}/{replyId}": {
      "delete": {
        "tags": [
          "Respostas automáticas"
        ],
        "summary": "Excluir resposta automática",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "replyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Regra excluída"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Regra não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Respostas automáticas"
        ],
        "summary": "Atualizar resposta automática",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "replyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "keyword"
                ],
                "properties": {
                  "keyword": {
                    "type": "string"
                  },
                  "response": {
                    "type": "string"
                  },
                  "isMedia": {
                    "type": "boolean"
                  },
                  "mediaUrl": {
                    "type": "string"
                  },
                  "mediaType": {
                    "type": "string",
                    "enum": [
                      "image",
                      "video",
                      "document",
                      "audio"
                    ]
                  },
                  "triggerType": {
                    "type": "string",
                    "enum": [
                      "ALL",
                      "GROUP",
                      "PRIVATE"
                    ]
                  }
                }
              },
              "example": {
                "keyword": "hello",
                "response": "Hi there! How can I help?",
                "matchType": "EXACT",
                "triggerType": "ALL",
                "isMedia": false,
                "mediaUrl": null,
                "mediaType": null
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Regra atualizada"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Regra não encontrada"
          }
        }
      }
    },
...
  },
...
}